pyproject.toml
[project]
name = "multiagent-ds"
version = "0.1.0"
description = "CSV-only Phase 1: toward replacing a data scientist"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.1",
  "numpy>=1.26",
  "duckdb>=0.9",
  "scikit-learn>=1.3",
  "matplotlib>=3.8",
  "pyyaml>=6.0.1",
  "rich>=13.7",
  "pyarrow>=21.0.0",
  "fastparquet>=2024.11.0",
]

[project.optional-dependencies]
dev = [
  "ruff==0.5.7",
]
